Re-implemented distinct taxon count

+++ lists-service/src/main/java/au/org/ala/listsapi/model/SpeciesList.java
@@ -32,7 +32,7 @@ public class SpeciesList {
   private List<String> facetList;
   private String doi;
   private Integer rowCount;
-  private Integer distinctMatchCount;
+  private Long distinctMatchCount;
   private String authority;
   private String category;
   private String region;
